Fix `CircuitBreakerClient` to return a `HttpResponse` immediately without waiting for headers or trailers (#3631)
Motivation:

Currently, `CircuitBreakerClient` can wait for `RequestLogProperty.RESPONSE_TRAILERS`
depending on `CircuitBreakerRule`.
https://github.com/line/armeria/blob/7bf1873fa71766f1040a6581ed1dd0e3687b51a8/core/src/main/java/com/linecorp/armeria/client/circuitbreaker/CircuitBreakerClient.java#L295-L298

The `responseFuture` can be completed only when a trailers is received
from remote peer. Even though a trailers was sent by remote peer, it
cannot be passed to `RequestLog`. As a consequence, the `resposneFuture`
cannot be completed until a `ResponseTimeoutException` is rased.
Since `InboundTrafficController` blocks to read data from channel buffer
when messages of `DecodedHttpResponse` is not comsumed by a `Subscriber`.
https://github.com/line/armeria/blob/6f21290a3e6e3bb2af22d8a9061d472dcabda545/core/src/main/java/com/linecorp/armeria/client/DecodedHttpResponse.java#L57-L59
https://github.com/line/armeria/blob/9061e241961c806951dd855be65fd871131a1835/core/src/main/java/com/linecorp/armeria/internal/common/InboundTrafficController.java#L72-L75

Modifications:

- For `CircuitBreakerRule`
  - Immediately return the orignal `HttpResponse` and report a circuit
    breaker result asynchronously.
- For `CircuitBreakerRuleWithContent`
  - Duplicate the origal `HttpResponse` and return the duplicated
  result immediately

Result:

You no longer see a `ResponseTimeoutException` with
`CircuitBreakerClient` when a large number of messages are received.