Provide tabs under each section instead for the whole procedure
Provide tabs for retrieving a single post and multiple posts, instead of dividing the whole page into two tabs. Extra work is required to render content Toc on the side if the whole page is divided into two. Providing tabs under each section may be bothersome to readers for having to click a tab each time they want to see the instruction for getting multiple posts.