Exclude flaky tests using annotation
Motivation:

It is hassle to add `assumeThat(..)` for excluding flaky tests.
It is also not a good way to exlude a whole class or its subclass.
For convenient sake, we can use JUnit5's `@Tag` annotation and filter in or out
using `includeTags` and `excludeTags`.

Modifications:

- Add `@FlakyTest` annotation to exclude a flaky test method or class
  from smoke test.
- Mark `GracefulShutdownIntegrationTest` as a flaky test.

Result:

Easy to maintain flaky tests