Fix to call `onRemoval()` immediately when using `StreamMessage.collect()`.
Motivation:

`DecodedHttp{Request,Response}` remove an object from the inbound
traffic when `onRemoval()` is called. The current implementation of
`StreamMessage.collect()` in `DefaultStreamMessage` calls `onRemoval()`
when a stream is closed via `close()`. That means the inbound traffic
will be removed only when the messages of a stream is fully received.

Modifications:

- Immediately drain the objects from the `queue` of
  `DefaultStreamMessage` whenever they are written to the `queue`
  when `collect()` is called so that `onRemoval()` is called right away.

Result:

You no longer see a `ResponseTimeoutException` when collecting a large number of
message using `Http{Request,Response}.aggreagte()`.