Move tutorial files to sub-directory to distinguish tutorial type
Move tutorial files to /rest/blog directory, as suggested by Trustin. To draw up the menu pane navigation properly, query the relative directory of each file and check the relative directory value in generating links.

Only change the template for tutorials, so the original docs do not query the relative directory. However, since MDXLayout is re-used by tutorial, the check for relative directory is applied for both docs and tutorials.