Add `JacksonObjectMapperProvider`
Motivation:

Currently, `JacksonModuleProvider` can customize the Jackson's modules
of the default Jackson `ObjectMapper`.
However it cannot configure the `ObjectMapper` such as
`DeserializationFeature` and `SerializationFeature`.
As the default `ObjectMapper` is widely used many places such as `HttpRequest`,
`HttpResponse`, `WebClient` and so on, it should be convenient to
configure the default `ObjectMapper` with a SPI.

Modifications:

- Add `JacksonObjectMapperProvider` that returns an `ObjectMapper`
  provided via SPI
- Breaking) `JacksonModuleProvider` has been removed in favor of
  `JacksonObjectMapperProvider`
- Enable `DeserializationFeature.FAIL_ON_NULL_CREATOR_PROPERTIES` as
  a sentible default option for Scala because `Option` is commonly
  preferred rather than a null value.

Result:

- You can now specify the default `ObjectMapper` using
`JacksonObjectMapperProvider`
- You no longer see null values when a case class is decoded from a
  JSON in Scala.
  - 404 Bad Request is returned instead.