Add `CompletableHttpResponse` whose actual `HttpResponse` is lazily completed
Motivation:

If a request is canceled, a `CompletableFuture` completes
a `HttpResponse` will be canceled. #3691
If the `CompletableFuture` was canceled and a `HttpResponse` tries to
complete the future later, the `HttpResponse` cannot be set and it would
be leaked.
It would be nice to abort a `HttpResponse` if a `CompletableFuture` was
already completed.

Modifications:

- Add `CompletableHttpResponse` that delegate the actual `HttpResponse`
  completed later.
- Use `CompletableHttpResponse` instead of `DeferredHttpResponse` when
  a `HttpResponse` is created with `HttpResponse.from(CompletableFuture<?>)`
- Deprecate `HttpResponse.from(CompletableStage, EventExecutor)`
  - No longer need to specify `EventExecutor` when creating a
    `HttpResponse from `CompletionStage`.

Result:

You no longer see leaks when a returned `CompletableFuture<HttpResponse>` was already
completed before a `HttpResponse` is completed.