Should not cancel `CompletableFuture` returned in annotated service (#3796)
Motivation:

If a request is canceled, a `CompletableFuture` which completes
a `HttpResponse` will be canceled. #3691
If the `CompletableFuture` was already canceled and a `HttpResponse`
tries to complete the future later, the `HttpResponse` cannot be set
and it would be leaked.

Actually, the cancellation of a `CompletableFuture` is not propagated to
the upstream futures. Furthermore, there is no event listener when a
value is failed to complete like `Flux.doOnCancel()`. So we decided not to cancel
`CompletableFuture` and abort the value when a request was already canceled.
But Kotlin Coroutines and Publisher are still canceled like now.

Modifications:

- Revert code in #3691 that proactively cancels a `CompletableFuture` in
  annotated service.
- Move `StreamMessageUtil` to the internal package. 

Result:

You no longer see leaks when a returned `CompletableFuture<HttpResponse>`
was already completed before a `HttpResponse` is completed.