Align the call order of `Server.Listner` of gRPC service with the upstream
Motivation:

If a call is exceptionally completed with non-OK status,
`ArmeriaServerCall` invokes `Listener.onCancel()` instead of
`Listener.onComplete()`.
https://github.com/line/armeria/blob/40990d3ab03ef0cecab446f011d117e8e0490d7a/grpc/src/main/java/com/linecorp/armeria/server/grpc/ArmeriaServerCall.java#L564-L577
The behavior is different of `gRPC-Java` that invokes
`Listener.onComplete()` regardless of the `Code` of `Status`.
The upstream only calls `Listener.onCancel()` if a transport is
closed before fully sending responses or a unary response is completed
without a response.

Modifications:

- Invoke `onReady()` only if a response is stream.
- Invoke `onComplete()` if a call is finished by a user.
- Invoke `onCancel()` if a call is finished by a client timeout
  scheduler.
- Invoke `onCancel()` if a call is finished without a response with a
  unary response method.
- Remove the invocation of `onReady()` after `invokeHalfClose()`.
  - The upstream calls `ServerCallStreamObserverImpl.onReadyHandler()
    that is different from `onReady()`

Result:

`GrpcService`s now correctly invoke `Server.Listener` in order.