Automatically decompress a compress file with `FileService`
Motivation:

A `FileService` serve compressed files depending on the accept-encoding
when `FileServiceBuilder.serveCompressedFiles(boolean)` is enabled.
If a compressed file such as `foo.json.gz` only exists, a request for
`foo.json` would be failed.
The size of uncompressed files are relatively big than compressed one.
Some users might not want to package the uncompressed files to a JAR file.

Modifications:

- Add `autoDecompress` option to `FileServiceBuilder`.
- Decompress a compressed file if a client want to a uncompressed content.
- Add `StaticHttpDecodedResponse` that decodes `HttpObject`s using
  a predefined `StreamDecoder`.
- Fix `FileService` to cache uncompressed files.
- Move some package-private classes to the internal packages.

Result:

- You can now decompress compressed files when use `FileService`
- Fixes #3804