Fix a bug where a client sends requests using a closing connection.
Motivation:

If a server sends a GOAWAY frame with the last stream
identifier set to 2^31-1 and a NO_ERROR code, a client should not
initiate further requests.
https://datatracker.ietf.org/doc/html/rfc7540#section-6.8
However, a client does not immediately removes the connection from the
connection pool. The connection will be removed from the connection pool
when the connection is about to be closed.
Hence, new requests sent to the server could be finished with
`ClosedSessionException` in the middle of processing.

Modifications:

- Deactivate a `Channel` from the connection pool as soon as a GOAWAY is received.
- Fix a bug where a GOAWAY is not flushed immediately.

Result:

You no longer see `ClosedSessionException` when a server gracefully
shut down a HTTP/2 connection.