Exclude `%2F(/)` from decoding of percents in a request path.
Motivation:

Currently, `%2F` and `%2f` are converted into `/` while decoding
request paths. It might route a path to an unexpected service if users
really want to send `%2F` as data itself.
Furthermore, other specifications such as gRPC transcoding do not
expect to decode `%2F` as `/`.
https://github.com/googleapis/googleapis/blob/02710fa0ea5312d79d7fb986c9c9823fb41049a9/google/api/http.proto#L257-L258
It should be better to exclude `%2F` for less confusion and better
interop with other echo systems.

Modifications:

- Exclude `%2F` and `%2f` from decoding percent-encoded characters in paths.

Result:

`%2F` and `%2f` are no longer converted to `/` when decoding a request path.