Ignore `management.server.port` if Actuator integration is not configured
Motivation:

A server created by Spring autoconfiguration listens to
`management.server.port` even if Actuator for Armeria is not configured.
That causes a port collision if non-Armeria Actuator module tries to
bind to the port.

Modifications:

- Created a virtual host for `management.server.port` only when
  `ArmeriaSpringActuatorAutoConfiguration` is configured.

Result:

`management.server.port` is now correctly activated only when
Armeria Actuator is configured.