Fix a bug where a gRPC server call is started before a context intialization
Motivation:

If a `DynamicEndpointGroup` is initially completed with empty endpoints,
an `IllegalStateException` is raised like the following:
```java
java.lang.IllegalStateException: Should call init(endpoint) before invoking this method.
    at com.google.common.base.Preconditions.checkState(Preconditions.java:502)
    at com.linecorp.armeria.client.DefaultClientRequestContext.eventLoop(DefaultClientRequestContext.java:575)
    at com.linecorp.armeria.internal.client.grpc.ArmeriaClientCall.start(ArmeriaClientCall.java:248)
    at io.grpc.stub.ClientCalls.startCall(ClientCalls.java:332)
```

A `ClientRequestContext` should be initialized sucessfully later when a non-empty
endpoints are received within a connection timeout.
https://github.com/line/armeria/blob/d78b52f3bc62ed1c4f5037f092ea1f514156d880/core/src/main/java/com/linecorp/armeria/client/DefaultClientRequestContext.java#L311
Otherwise, it would be failed with `EmptyEndpointGroupException`.

Modifications:

- Mark a EndpointGroup was successfully initialized only when the
  current endpoints are non empty.

Result:

You no longer see `IllegalStateException` when a gRPC client call
starts with an empty endpoint.