Add a script that generates skeletal release notes
Motivation:

It takes a couple of hours to collect information of chages and wirte
them to the new release notes.
The added script will write a draft version of a release note to help maintainers
revise the release note and publish it easily.

Modifications:

- Add `release-note.ts` to `site` module.

Result:

We don't need to write a release note from scratch.
```
cd site/
npm run release-note <version>
```