Abort server startup if `ServerListener.serverStarting` throws an exception (#3947)
Motivation:

In our javadoc, it is stated that if `ServerListener#serverStarting` throws an exception `Server` startup is aborted.
https://github.com/line/armeria/blob/43bd08248ba86ab5942f781da33b11a42fe3b1a5/core/src/main/java/com/linecorp/armeria/server/ServerListener.java#L34-L37

However, this behavior currently isn't supported since the introduction of `StartStopSupport`.

Additionally, we have verified that this is still a valid use case. (i.e. users may want to do some validation before proceeding with server startup)

Modifications:

I think it is difficult to cleanly decouple the behavior of `StartStopSupport` and `Server` without breaking changes (see 580560de57c77755300460c3aeccd540366148bf for the attempt). I also think it is acceptable for `StartStopSupport` to support the behavior where an exception in `notifyStarting` fails startup.

- Modify `StartStopSupport#notifyListeners` to throw an exception if `state == State.Starting `
    - Eventually, the exception will be thrown before startup starts in the following line: https://github.com/line/armeria/blob/a1105519e920b00ee3bc848a5074d943925931ef/core/src/main/java/com/linecorp/armeria/common/util/StartStopSupport.java#L184
- Add `try..catch` statements which ignores the exception (since other states won't propagate the exception + there is no additional behavior required to restore from the exception)

Result:

- Closes #3935
- `StartStopSupport` now aborts startup if `notifyStarting` throws an exception.