Use React hooks for getting current path
Use the code offered by @minwoox. Reference: https://github.com/line/armeria/pull/3970#issuecomment-987587701

On top of the code offered, add a second parameter to  `useEffect()` and bring the content of the `useTutorialType()` a level up, because the linter saw this as an error, as the hook was being called inside a callback.