Resolve `WildcardType` to get upper bound (#4008)
Motivation:

When I use a wildcard type for parameter like as follows, `AnnotatedService` is failed to initialize due to `Unsupported or invalid parameter type: ?`:

```java
@Get("/wildcard")
public String wildcard(RequestContext ctx, @Param List<? extends Object> container) {
    return ctx.path();
}
```

It'd be ok for Java users since they may not use like the above. However, for Kotlin, it isn't when they use an interface as a generic type. IMHO, this is because Kotlin changes the type to the upper-bounded wildcard type. This is the reproducible code:
```kotlin
class AnnotatedService {
    @Get("/")
    fun foo(@Param("values") values: List<MyInterface>): String {
        return "Hello, $values!";
    }
}

interface MyInterface
```

Modifications:

- Resolve `WildcardType` to get the first upper bound when getting the raw type in annotated services' `@Param`s.

Result:

- Closes #4007
- Users can use upper-bounded wildcard type as a `@Param` in annotated services.
  - Especially Kotlin users can use interface as a generic type.