Allow reading a specific range of a file using `PathStreamMessage`
Motivation:

The [Range][1] HTTP request header indicates the part of a document that the server should return.
Currently, `StreamMessage.of(Path)` reads all content of a file from the beginning.
If `StreamMessage.of(Path)` is able to read a content from a specific
range, it would be useful to handle `Range` header.

Modifications:

- Add `start` and `end` parameters to `PathStreamMessage` to indicate
  the range of the file to read.

Result:

You can now read a range of a file using `StreamMessage.of(Path, start, end)`.
```java
Path partialContent = ...;
StreamMessage.of(partialContent, 100, 200);
```

[1]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Range