Fix not to use `Http{Response,Status}Exception` internally
Motivation:

An `HttpResponseException` has an `HttpResponse`. So the default
`ServerErrorHandler` extracts the `HttpResponse` from the exception.
As the response is wrapped by `HttpResponseException`,
decorators can not access the `HttpObject`s of the response in the
exception.

Users may find it hard to mutate the thrown `HttpResponse` because
recovery should be done first. So it would be better not to use
`HttpResponseException` or recover it before passing a response to
decorators.

Modifications:

- Add `HttpFile.ofRedirect(location)` to indicate a file in a different
  location.
- Recover an `Http{Response,Status}Exception thrown by
  `HealthCheckUpdateHandler`
- Update Javadoc in detail for `Http{Response,Status}Exception`
- Correctly propagate the cause of `Http{Response,Status}Exception` in
  `THttpService`

Result:

- You can now mutate a redirect response from `FileService` using
  `mapHeaders`
- You no longer see an `HttpResponseException` or
  an `HttpStatusException` from built-in services.
- Fixes #4056