fix: grpc doc generation failing for options and extensions (#4127)
Motivation:

gRPC doc generation fails when options and extensions were being used in proto files.

The test cases have been modified so the `test.proto` file contains an example option. Before changing `GrpcDocStringExtractor`, it wasn't able to produce documentation because of a duplicate key in the map. It was causing the whole Documentation Service to crash.

We might want to consider displaying `options` in documentation service in the future but this implementation un-blocks users who can't use the service because of the aforementioned bug.

The issue is described in #4123 in more detail.

Modifications:

- Ignore comments written on proto service options or extensions.

Result:

- Closes #4123
- The documentation service doesn't crash the whole Armeria service anymore when proto files have comments on service options or extensions.

Co-authored-by: jrhee17 <guins_j@guins.org>