Fix not to use `Http{Response,Status}Exception` internally (#4117)
Motivation:

An `HttpResponseException` has an `HttpResponse`. The thrown response
is extracted by the default `ServerErrorHandler` and converted to 
a normal `HttpResponse`. There are downsides to `HttpResponseException`.
As the response is wrapped by `HttpResponseException`,
decorators can not access the `HttpObject`s of the response in the
exception using `mapXXX`.

Users may find it hard to mutate the thrown `HttpResponse` because 
recovery should be done first. Therefore, it would be better not to use 
`HttpResponseException` internally or recover it before passing
the response to decorators so that users use `HttpResponse.mapXXX` to
transform a returned response in decorators.

Modifications:

- Add `HttpFile.ofRedirect(location)` to indicate a file in a different
  location.
- Recover an `Http{Response,Status}Exception` thrown by
  `HealthCheckUpdateHandler`
- Update Javadoc in detail for `Http{Response,Status}Exception`
- Correctly propagate the cause of `Http{Response,Status}Exception` in
  `THttpService`

Result:

- You can now mutate a redirect response from `FileService` using
  `mapHeaders`
- You no longer see an `HttpResponseException` or
  an `HttpStatusException` from built-in services.
- Fixes #4056