Handle request timeout correctly by comparing status codes (#4165)
Motivation:

#3060 had handled request timeout on http1 connections.
However, I had falsely assumed that netty `HttpResponseStatus` was a singleton and compared by reference.
It turns out that this isn't the case if the provided `reasonPhrase` doesn't match that of `netty`'s. (the reported reason phrase was `Request-Timeout` with a hyphen)

https://github.com/netty/netty/blob/824a0b803431c58bd8dfe25a30413d7ff9759e8f/codec-http/src/main/java/io/netty/handler/codec/http/HttpResponseStatus.java#L467

Modifications:

- Modify `ArmeriaHttpUtil#isRequestTimeoutResponse` to compare by status code.

Result: 

Now, even for custom reason phrases request timeout on h1 should be correctly handled.