Display JSON and protobuf request and response specification in `DocService`
Motivation:

A `@RequestObject` bean can be converted as a specification of an
annotated service. Neither response nor pojo request are shown in a
`DocService`. Only gRPC and Thrift docservice displays the full
speficiations of input and output type.

This PR aims to generalize a way to create a `StructInfo` from a range
of types by offering `NamedTypeInfoProvider` which can be dynamically
loaded via SPI or explicitely set using `DocServiceBuilder`.
A data type that depend on a specific library, such as protobuf,
Thrift, ScalaPB can extend this interface and inject custom
implementations.

Modifications:

- Designed `NamedTypeInfoProvider` interface to let users customize how to
  create a `StructInfo` from the given type descriptor.
  The following implmentations are adedd for the default behavior.
  - `JsonNamedTypeInfoProvider`
  - `ThriftNamedTypeInfoProvider`
  - `ProtobufNamedTypeInfoProvider`
  - `ScalaPbNamedTypeInfoProvider`
- Add `DocServiceBuilder.nameTypeInfoProvider()` in order to override
  the default `NamedTypeInfoProvider`s.
- Moved some code related to `StructInfo` in `*DocServicePlugin` to
  `*NamedTypeInfoProvider`

Result:

- You can now see the speficication of an annotated service maded of a JSON
  object or protobuf can be viewed in `DocService`.
- Fixes #4309