Add `UnaryServerCall` (#4192)
Motivation:

`ExchangeType` is introduced by #3956. A unary call is automatically
aggregated by `Http{1,2}RequestDecoder` and `HttpServerHandler`.
The optimized call path for non-stream requests showed better
performance. As a part of the performance of optimizations, we can add a
specialized `ServerCall` implementation for unary calls. Because
unary calls are widely used for general communication.
Even if the current implementation is not a significant improvement, we
can add further optimization based on this.

Modifications:

- Split `ArmeriaServerCall` into `UnaryServerCall` and
  `StreamingServerCall`.
  - `AbstractServerCall` is added to share common logic.
- Add `UnaryDecoderInput` as a variant of `StreamDecoderInput` which
  holds a single `ByteBuf`.
  - Renamed `ByteBufDecoderInput` into `ByteBufsDecoderInput` to
    differenciate from `UnaryDecoderInput`
- Add `UnaryMessageDeframer` that directly decodes an `HttpData` into
  `DeframedMessage`.
- Miscellaneous) Remove an unexpected string concatenation when validating
  `ResponseHeaders` in `RegularFixedHttpResponse`

Result:

Better performance for unary gRPC call.