Migrate `dependencies.yml` to Gradle's version catalogs
Motivation:

We can declare dependency versions using `dependencies.yml` and share it
across multiple projects. It is useful to manage the versions of various
dependencies. However, we can't define different versions of a module.
Because the unique keys in `dependencies.yml` are `moduleId:artifactId`.

This limitation makes it difficult to manage compatibility between
different versions of a library. For example, okhttp v4 is required to
test `it:okhttp` and okhttp v3 is required to run `retrofit2` module.

Gradle introduced a new way to [share dependency version between projects](https://docs.gradle.org/current/userguide/platforms.html).
The unique key of version catalogs is a user-defined alias so we can
define multiple versions of a module using different aliases.

Version catalogs does not allow extra properies on `libs.versions.toml`.
It only allow some properties for versions and modules.
`dependencies.yml` not only manages versions, but also manages metadata
for a module such as relocations, exclusions and javadoc links.

I wanted to take advantage of both. We need a version manager that
supports version catalogs based on aliases, and ability to add the
metadata. As a result, I implmented to add our own dependency machanism
using `dependencies.toml` that is fully compatible with Gradle's version
catalogs and understands relocations, exclusions and javadoc links
declearations.

Modifications:

- Migrate `dependencies.yml` to `dependencies.toml`
- TBU

Result:

- Fixes #4120