Fix a bug where `HealthEndpointGroup` raises `NoSuchElementException` while closing
Motivation:

`HealthEndpointGroup.allHealthyEnpoints()` is used to get all health
endpoints from `contextGroupChain`, which is called when either
healthness of an `Endpoint` is updated after an health check
https://github.com/line/armeria/blob/554d5963bef33162b1cb077077336de3ef3aadfb/core/src/main/java/com/linecorp/armeria/client/endpoint/healthcheck/HealthCheckedEndpointGroup.java#L252-L263
or when the first contextGroup is resolved.
https://github.com/line/armeria/blob/554d5963bef33162b1cb077077336de3ef3aadfb/core/src/main/java/com/linecorp/armeria/client/endpoint/healthcheck/HealthCheckedEndpointGroup.java#L171-L181

`allHealthyEndpoints`, in turn, assumes that at laeat one contextGroup is
available when it is called. However if an `Endpoint` is updated while
an `HealthEndpointGroup` is closing, `allHealthyEndpoints` can see an
empty `contextGroup` and the following error is warned.
```
Caused by: com.linecorp.armeria.common.util.CompositeException$ExceptionOverview: Multiple exceptions (2)
|-- java.util.NoSuchElementException: null
    at java.base/java.util.ArrayDeque.getLast(ArrayDeque.java:413)
|-- com.linecorp.armeria.common.stream.AbortedStreamException: null
    at com.linecorp.armeria.common.stream.AbortedStreamException.get(AbortedStreamException.java:39)
```

Modifications:

- Return an empty list if `contextGroupChain` is empty
- Do not update the healthness of an `Enpoint` if `HttpHealthChecker` is
  closing

Result:

You no longer see a `NoSuchElementException` while an
`HealthEndpointGroup` is closing.