Fix a bug where request duration is incorrectly recoded.
Motivation:

The start time of a request indicates when a server receives the first
headers of a request and it is recorded when `HttpServiceHandler`
receives the decoded request from `Http{1,2}RequestDecodeder`.
Previously, `Http{1,2}RequestDecoder` passes a request right after the
first headers is received. So it was correctly recoded as expected.
However, after `ExchangeType` has been introduced, a fully aggregated
request is passed to `HttpServerHandler` if `ExchangeType.UNARY` or
`ExchangeType.RESPONSE_STREAMING` is set.

Modifications:

- Recode `requestStartTimeNanos` and `requestStartTimeMicros` when
  request headers is received in `Http{1,2}RequestDecoder`

Result:

- `requestStartTimeNanos` and `requestStartTimeMicros` are now correctly
  recoded when `RequestHeaders` is received.
- Fixes #4365