Retry when Spring tests fails to start a `Server` with a random port obtained in advance
Motivation:

If a random port (0) for internal services or managed services is
specified in Spring configuration, the Spring integration module
allocate available ports early and use them to configure additional
service later.
https://github.com/line/armeria/blob/948763acc7911c951e37dbd35132edd253ea3934/spring/boot2-actuator-autoconfigure/src/main/java/com/linecorp/armeria/spring/actuate/ArmeriaSpringActuatorAutoConfiguration.java#L258-L260
This is not a problem in production mode.
However, in testing situation, many test servers are started with a
random port and the available ports can be acquired by them. #4391 #4329

Modifications:

- Add `RetryableArmeriaServerGracefulShutdownLifecycle` to retry a
server with a backoff at most N times.

Result:

- This is not a perfect answer for the failure but it might reduce the
  number of failures.
- Fixes #4391 #4329