Fix the wrong Guava `failureaccess` version
Motivation:

The dependency version of `com.google.guava:failureaccess` should be
`1.0.1` but Guava's version (`31.1-jre`) is set.
Actually, it does not cause any problems because only the metadata such
`exclusions` and `relocations` are used to shade Guava.
Anyway, it would be better to use the correct value for being future-proof.

Modifications:

- Set `com.google.guava:failureaccess` to 1.0.1

Result:

Fix a migration mistake.