Rename `newResponseConverterFunction` to `createResponseConverterFunction` in `ResponseConverterFunctionProvider`
Motivation:

`DelegatingResponseConverterFunctionProvider` and
`RequestConverterFunctionProvider` uses `create` as a prefix of the API.
But `ResponseConverterFunctionProvider` uses `new` as a prefix.
A consistant naming for the similar APIs can give better developer
experience.

Modifications:

- Rename `newResponseConverterFunction` to `createResponseConverterFunction`
  in `ResponseConverterFunctionProvider`

Result:

Better developer experience.
Note that this is not an additional breaking change because it has not
been deployed yet.